+++ trixnity-messenger-compose-app/src/commonMain/kotlin/de/connect2x/messenger/messengerConfiguration.kt
@@ -27,9 +27,10 @@ fun messengerConfiguration(): MatrixMultiMessengerConfiguration.() -> Unit = {
         platformGetSystemLangModule(),
         module {
             single<Languages> { DefaultLanguages }
-            single<I18n> { object : I18n(get(), get(), get()) {} }
+            single<I18n> { object : I18n(get(), get(), get(), get()) {} }
         },
     )
+    multiProfile = false
 
     // MatrixMultiMessengerConfiguration flavors
     when (BuildConfig.flavor) {