+++ trixnity-messenger-compose-view/src/webMain/kotlin/de/connect2x/messenger/compose/view/Notifications.kt
@@ -1,17 +1,16 @@
-package de.connect2x.messenger.desktop
+package de.connect2x.messenger.compose.view
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.window.Notification
-import androidx.compose.ui.window.TrayState
-import de.connect2x.messenger.compose.view.DI
-import de.connect2x.messenger.compose.view.IsFocused
-import de.connect2x.messenger.compose.view.files.imageBitmapFromBytes
-import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
+import de.connect2x.sysnotify.Notification
+import de.connect2x.sysnotify.NotificationHandler
+import de.connect2x.sysnotify.NotificationIcon
+import de.connect2x.sysnotify.create
 import de.connect2x.trixnity.messenger.MatrixClients
 import de.connect2x.trixnity.messenger.MatrixMessenger
 import de.connect2x.trixnity.messenger.MatrixMessengerAccountSettings
+import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
 import de.connect2x.trixnity.messenger.MatrixMessengerSettingsHolder
 import de.connect2x.trixnity.messenger.platformNotifications
 import de.connect2x.trixnity.messenger.viewmodel.util.RoomName
@@ -19,7 +18,10 @@ import de.connect2x.trixnity.messenger.viewmodel.util.avatarSize
 import de.connect2x.trixnity.messenger.viewmodel.util.scopedCollectLatest
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.conflate
+import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import net.folivo.trixnity.client.MatrixClient
@@ -36,24 +38,44 @@ import net.folivo.trixnity.core.model.events.m.room.Membership
 import net.folivo.trixnity.core.model.events.m.room.RoomMessageEventContent
 import net.folivo.trixnity.core.model.events.roomIdOrNull
 import net.folivo.trixnity.core.model.events.senderOrNull
-import net.folivo.trixnity.core.model.push.PushAction
 import net.folivo.trixnity.utils.toByteArray
-import javax.sound.sampled.AudioSystem
 
 private val log = KotlinLogging.logger { }
 
 @Composable
 fun Notifications(
     matrixMessenger: MatrixMessenger,
-    trayState: TrayState,
 ) {
     val i18n = DI.get<I18nView>()
+    val config = matrixMessenger.di.get<MatrixMessengerConfiguration>()
+    val notificationHandler = NotificationHandler.create(
+        name = "${config.appName} Notifications",
+        id = "${config.packageName}.${config.appName}.notification",
+    )
+
+    LaunchedEffect(Unit) {
+        matrixMessenger.di.get<MatrixMessengerSettingsHolder>()
+            .map { it.base.accounts }
+            .distinctUntilChanged()
+            .conflate()
+            .collect { settings ->
+                val anyNotificationsEnabled =
+                    settings.any { (_, settings) -> settings.base.notificationsEnabled }
+                withContext(Dispatchers.Main) {
+                    if (anyNotificationsEnabled) {
+                        log.debug { "Notifications are enabled for active messenger, requesting permissions" }
+                        notificationHandler.requestPermissions()
+                    }
+                }
+            }
+    }
 
     val windowIsFocused = IsFocused.current
     LaunchedEffect(windowIsFocused) {
         withContext(Dispatchers.Default) {
+            log.debug { "window is focused: $windowIsFocused" }
             val roomNameComputation = matrixMessenger.di.get<RoomName>()
-            whenSyncIsRunning(matrixMessenger, windowIsFocused, roomNameComputation, trayState, i18n)
+            whenSyncIsRunning(matrixMessenger, windowIsFocused, roomNameComputation, notificationHandler, i18n)
         }
     }
 }
@@ -62,7 +84,7 @@ private suspend fun whenSyncIsRunning(
     matrixMessenger: MatrixMessenger,
     windowIsFocused: Boolean,
     roomNameComputation: RoomName,
-    trayState: TrayState,
+    notificationHandler: NotificationHandler,
     i18n: I18nView
 ) {
     val settings = matrixMessenger.di.get<MatrixMessengerSettingsHolder>()
@@ -74,19 +96,7 @@ private suspend fun whenSyncIsRunning(
                     log.debug { "windowIsFocused: $windowIsFocused" }
                     if (windowIsFocused.not() && currentSettings.base.notificationsEnabled) {
                         log.debug { "received notification for event ${notification.event.idOrNull}" }
-                        if (currentSettings.platformNotifications.notificationsPlaySound &&
-                            notification.actions.any { it is PushAction.SetSoundTweak }
-                        ) {
-                            withContext(Dispatchers.IO) {
-                                val audioStream =
-                                    AudioSystem.getAudioInputStream(
-                                        MessengerTrayIcon::class.java.getResourceAsStream("/ding.wav")?.buffered()
-                                    )
-                                val clip = AudioSystem.getClip()
-                                clip.open(audioStream)
-                                clip.start()
-                            }
-                        }
+                        // sound from web?
                         if (currentSettings.platformNotifications.notificationsShowPopup) {
                             val room = notification.event.roomIdOrNull?.let { matrixClient.room.getById(it).first() }
                             val isDirect = room?.isDirect ?: false
@@ -100,7 +110,9 @@ private suspend fun whenSyncIsRunning(
                                 isDirect,
                                 roomName,
                                 i18n
-                            )?.let { trayState.sendNotification(it) }
+                            )?.let {
+                                notificationHandler.push(it)
+                            }
                         }
                     }
                 }
@@ -127,14 +139,11 @@ private suspend fun displayNotification(
         }
 
         if (message != null) {
-            val (username, _) = event.senderOrNull?.let { sender ->
+            val (username, imageInBytes) = event.senderOrNull?.let { sender ->
                 val user = matrixClient.user.getById(roomId, sender).first()
                 val image = user?.avatarUrl?.let { avatarUrl ->
                     matrixClient.media.getThumbnail(avatarUrl, avatarSize().toLong(), avatarSize().toLong())
-                }?.map { it.toByteArray() }
-                    ?.map { bytes ->
-                        imageBitmapFromBytes(bytes)
-                    }?.getOrNull()
+                }?.map { it.toByteArray() }?.getOrNull()
                 user?.name to image
             } ?: (null to null)
 
@@ -143,7 +152,10 @@ private suspend fun displayNotification(
 
             log.debug { "notification will appear" }
 
-            return Notification(title, text)
+            return Notification(
+                title = title,
+                description = text,
+                icon = imageInBytes?.let { NotificationIcon(it, avatarSize(), avatarSize()) })
         }
     } ?: log.warn { "cannot find roomId for event ${event.idOrNull}" }
     return null