+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/util/constants.kt
@@ -1,5 +1,3 @@
 package de.connect2x.trixnity.messenger.viewmodel.util
 
-fun timezone() = "CET" // TODO configurable!
-
 fun avatarSize() = 36